[parts 1 and 2] readability, typography, grammar
@Rune23: I plan to go through the remaining parts this evening and do
more of the same, so let me know if there's anything you don't like.

One thing I'm doing for example is change "Dai" to "dai".

Other than that I'm mostly reading the text and rewriting or changing
the parts I find difficult to read, so hopefully I'm not changing the
meaning of the text in any major ways.